this looks like it's working end-to-end now:
can clean the main sctp4j project and rebuild and a test in the 'sctp'
project is able to load the native binary correctly (currently some
hard-coding around this on mac)

there is now a 'disconnect' in the process.  the native jni library will
be built for the current platform (and put in targets/) but the bundled
native libraries are in src/main/resources.  so the libs should be built
per-platform and then copied there manually.