misc fixes
main one was that i was returning the wrong value from one of the
methods invoked by the sctp lib (0 instead of -1) to show error, which
was causing issues.  also have a hacky test in here (should be cleaned
up) which does a connection between a client and server and verifies
data can be sent.

there's also some other misc cleanup here: the new socket type added as
SctpSocket2 for now, and the jni library has no knowledge of the sctp
socket, it just invokes handlers for incoming/outgoing data that Sctp4j
installs...all the java abstraction logic is in Sctp4j now, so the jni
wrapper does only the wrapping logic and gets packets in/out.