Multiple changes:
1) Get rid of the dummy port thing and have the callers pass it in
instead (to both init and create socket).  Even though we don't really
use these, and our API changes things from what usrsctp does (e.g. we
pass the local port to createSocket, even though usrsctp_socket doesn't
need one because we hide the call to bind later (when doing connect)
which does want a local port) i think it's less confusing to be more
explicit about the ports

2) split out the client socket and server socket logic into separate
classes

3) fix a bug (and the example) which wasn't doing the accept properly:
client->server data can work without doing this, but not in the other
direction: this is because accept will return a new socket for the
accepted connection, and this is the one that must be used to send data
(though the wrapper doesn't actually return a new socket, it just
overwrites the socket pointer in the wrapper to be the new socket, since
we only ever accept one connection)