ref: Removes getPtrFromSocket()
As far as I can see this change is not functional. The ptr of a
SctpSocket is never changed, except when it is closed (at which point
it is also removed from the map). The pointer in the map is also not
updated anywhere.

I don't fully understand the comment in SctpServerSocket#accept about
the stack replacing a pointer, but I can not find anything in the native
code which changes the value of the 'ptr' field in java's SctpSocket.