fix: Fixes a deadlock.
Removes synchronization on SctpSocket, since it is called from the
usrsctp stack and is known to cause deadlocks. Brings back the
locking scheme from libjitsi[0,1] which assures that close is called
exactly once and deters calling it until all other operations
have completed, but avoids holding a lock while performing operations
(either in the SCTP stack, or in the application).

[0] https://github.com/jitsi/libjitsi/commit/cc053a7faa9acb99738ef6dced27b8515f21c3cf
[1] https://github.com/jitsi/libjitsi/commit/09c18e90f4b35a63adca8009cf7ba24ebb61f154